requirements.txt
# Pinned dependencies for reproducible builds
# Last updated: 2026-01-07 (v18.0 ML Embeddings)

playwright==1.57.0
playwright-stealth==2.0.0
pandas==2.3.3
duckdb==1.4.3
streamlit==1.52.2
altair==6.0.0
plotly==6.5.0
tqdm==4.67.1
lxml==6.0.2
PyYAML==6.0.3
Jinja2==3.1.6
google-genai>=0.1.0  # Modern Gemini SDK for weekly market insights
python-dotenv>=1.0.0
pytest

# Optional: ML-based role classification (420MB model download)
# Install with: pip install sentence-transformers
sentence-transformers>=2.2.0  # Optional - falls back to keyword matching if not installed
